requirements.txt
unstructured
unstructured[pdf]
openai 
lancedb
streamlit
streamlit-pdf-viewer==0.0.18
fastapi
uvicorn
python-dotenv